requirements.txt
appdirs==1.4.0
cffi==1.9.1
cryptography==1.7.2
Django==1.10.5
djangorestframework==3.5.4
idna==2.3
jwt==0.5.2
packaging==16.8
psycopg2==2.6.2
pyasn1==0.2.3
pycparser==2.17
pyparsing==2.1.10
six==1.10.0
